Make tests more resilient to updates in vulnerability data.
This should cause tests to fail when the format changes, but prevent
failures simply because a previously-not-known-to-be-vulnerable
version had a vulnerability discovered in it.

The caveat is that the `secure` bundle will still kvetch when one of
its gems comes up vulnerable, since it's expected to have ZERO
matches.  In that case, the relevant Gemfile/Gemfile.lock need to be
updated.