Using the ctime, or even mtime of the vendored directory is absolutely wrong.
Let's say you package up a version of the vendored repo with its last commit
being from time T=1.

You release your gem at time T=2.

The upstream vulnerability repo is updated at T=3.

A user installs the gem at T=4.

A user installs a local bundle at T=5.

Using the ctime of the vendored copy means that it will be comparing a
vendored copy with a timestamp of T=5 (but whose last commit is from T=1)
to a local copy with a last commit of T=2.

The code will thus use an OLDER vendored repo until such time as the vuln
database carries a commit newer than the time at which the user installed
the gem, and updates his or her local cache.