Move file reading responsibility to the CLI
This means Scanner is a simple wrapper for LockfileParser and accepts
the same arguments (content in the form of a string, not a directory path).

This also means that Scanner is able to be called with content from
anywhere (for example, github's content API) instead of a hardcoded file
path and makes it more useful as a library other projects can depend on.