Fix example for bundle without advisories
The update to the vendored ruby-advisory-db in 627805fa brought in
vulnerabilities for Rails 3.2, which was used by the example for a
bundle without advisories. This commit updates the example to use a
version of Rails without advisories, while also updating the vendored DB
to avoid some bugs in the version specifications.