fix: don't include test files in gem pkg
Spec files include Gemfiles with known vulnerabilities in them (on
purpose), but if scanned by a naive scanner (that is looking for any
Gemfile in the system), this can cause false positives.

In addition to not including this gem in production systems, we can also
just remove the specs from being included in the gem package in the
first place.