Add env to enable custom repository for ruby-advisory-db
Currently the advisories are always pulled from the official GitHub repository which always requires an internet connection. If you CI agent is not connected to the internet, it is currently not possible to update the database. With the introduction of an optional env variable, user can set their internal mirror instead.